pyproject.toml
[project]
name = "youtube-automation"
version = "0.1.0"
description = "Faceless YouTube channel automation - infographic & animation based video pipeline"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.10",
    "bar-chart-race>=0.2",
    "pandas>=2.2",
    "ffmpeg-python>=0.2",
    "moviepy>=2.0",
    "pillow>=10.0",
    "anthropic>=0.40",
    "google-api-python-client>=2.0",
    "google-auth-oauthlib>=1.0",
    "apscheduler>=3.10",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "requests>=2.31",
]

[project.optional-dependencies]
tts = ["elevenlabs>=1.0"]
whisper = ["faster-whisper>=1.0"]
dev = ["pytest>=8.0", "ruff>=0.8"]

[project.scripts]
yt-auto = "pipeline.main:cli"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
